Upgrade to rust 1.36 and use arrays for standing_orders instead of vectors (#155)
* get_standing_orders_of_slot will always return models::NUM_OF_ACCOUNT entries

* incorporated all inprovement suggestions, just refactoring

* using unsafe to initialize array, as Standingorder does not implement Copy trait

* clean up, sorry

* indentation

* uses array-macro to initialize the arrays

* test

* changing to newer rust version 1.36

* test only

* another test to make rust run on travis server

* just testing travis

* reverting all unnecessary changes

* updating rust to 1.36 in docker images

* making clippy work: see here: https://github.com/rust-lang/rust-clippy/issues/3874

* stupid testing as I can not yet test it locally

* worst testing

* clean up

* just another test

* one more test

* final test

* finally, Felix found the solution

+++ driver/src/models/mod.rs
@@ -13,7 +13,7 @@ use web3::types::H256;
 
 //ToDo: get variables from database
 pub const TOKENS: u8 = 30;
-pub const NUM_RESERVED_ACCOUNTS: u8 = 50;
+pub const NUM_RESERVED_ACCOUNTS: usize = 50;
 pub const DB_NAME: &str = "dfusion2";
 
 pub trait RollingHashable {
